Unlimited decorations
Decoration combi mesh count scales with requested usage
