+++ Orbis/Orbis.csproj
@@ -117,6 +117,7 @@
     </Compile>
     <Compile Include="Orbis.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
+    <Compile Include="Rendering\DecorationData.cs" />
     <Compile Include="UI\Elements\RelativeElement.cs" />
     <Compile Include="UI\IUpdatableElement.cs" />
     <Compile Include="UI\Elements\RelativeText.cs" />
